+++ src/compas_timber/_fabrication/__init__.py
@@ -18,6 +18,14 @@ from .dovetail_mortise import DovetailMortise
 from .dovetail_mortise import DovetailMortiseParams
 from .lap import Lap
 from .lap import LapParams
+from .french_ridge_lap import FrenchRidgeLap
+from .french_ridge_lap import FrenchRidgeLapParams
+from .tenon import Tenon
+from .tenon import TenonParams
+from .mortise import Mortise
+from .mortise import MortiseParams
+from .slot import Slot
+from .slot import SlotParams
 
 
 __all__ = [
@@ -39,4 +47,12 @@ __all__ = [
     "DovetailMortiseParams",
     "Lap",
     "LapParams",
+    "FrenchRidgeLap",
+    "FrenchRidgeLapParams",
+    "Tenon",
+    "TenonParams",
+    "Mortise",
+    "MortiseParams",
+    "Slot",
+    "SlotParams",
 ]