+++ src/compas_timber/connections/__init__.py
@@ -1,9 +1,9 @@
-from .french_ridge_lap import FrenchRidgeLapJoint
-from .joint import BeamJoinningError
+from compas_timber.errors import BeamJoinningError
 from .joint import Joint
 from .l_butt import LButtJoint
 from .l_halflap import LHalfLapJoint
 from .l_miter import LMiterJoint
+from .l_french_ridge_lap import LFrenchRidgeLapJoint
 from .lap_joint import LapJoint
 from .null_joint import NullJoint
 from .solver import ConnectionSolver
@@ -15,6 +15,7 @@ from .t_birdsmouth import TBirdsmouthJoint
 from .t_halflap import THalfLapJoint
 from .x_halflap import XHalfLapJoint
 from .t_dovetail import TDovetailJoint
+from .t_tenon_mortise import TenonMortiseJoint
 from .ball_node import BallNodeJoint
 
 __all__ = [
@@ -31,10 +32,11 @@ __all__ = [
     "THalfLapJoint",
     "LHalfLapJoint",
     "NullJoint",
-    "FrenchRidgeLapJoint",
+    "LFrenchRidgeLapJoint",
     "JointTopology",
     "ConnectionSolver",
     "find_neighboring_beams",
     "TDovetailJoint",
     "BallNodeJoint",
+    "TenonMortiseJoint",
 ]