+++ src/compas_timber/connections/joint.py
@@ -9,33 +9,6 @@ from compas_model.interactions import Interaction
 from .solver import JointTopology
 
 
-class BeamJoinningError(Exception):
-    """Indicates that an error has occurred while trying to join two or more beams.
-
-    This error should indicate that an error has occurred while calculating the features which
-    should be applied by this joint.
-
-    Attributes
-    ----------
-    beams : list(:class:`~compas_timber.parts.Beam`)
-        The beams that were supposed to be joined.
-    debug_geometries : list(:class:`~compas.geometry.Geometry`)
-        A list of geometries that can be used to visualize the error.
-    debug_info : str
-        A string containing debug information about the error.
-    joint : :class:`~compas_timber.connections.Joint`
-        The joint that was supposed to join the beams.
-
-    """
-
-    def __init__(self, beams, joint, debug_info=None, debug_geometries=None):
-        super(BeamJoinningError, self).__init__()
-        self.beams = beams
-        self.joint = joint
-        self.debug_info = debug_info
-        self.debug_geometries = debug_geometries or []
-
-
 class Joint(Interaction):
     """Base class for a joint connecting two beams.
 
@@ -108,6 +81,18 @@ class Joint(Interaction):
         """
         pass
 
+    def check_elements_compatibility(self):
+        """Checks if the beams are compatible for the creation of the joint.
+        This is optional and should only be implemented by joints that require it.
+
+        Raises
+        ------
+        :class:`~compas_timber.connections.BeamJoinningError`
+            Should be raised whenever the elements did not comply with the requirements of the joint.
+
+        """
+        pass
+
     def restore_beams_from_keys(self, model):
         """Restores the reference to the beams associate with this joint.
 
@@ -127,7 +112,7 @@ class Joint(Interaction):
         raise NotImplementedError
 
     @classmethod
-    def create(cls, model, *beams, **kwargs):
+    def create(cls, model, *elements, **kwargs):
         """Creates an instance of this joint and creates the new connection in `model`.
 
         `beams` are expected to have been added to `model` before calling this method.
@@ -151,7 +136,7 @@ class Joint(Interaction):
 
         """
 
-        joint = cls(*beams, **kwargs)
+        joint = cls(*elements, **kwargs)
         model.add_joint(joint)
         return joint