+++ src/compas_timber/elements/fastener.py
@@ -1,7 +1,5 @@
 # TODO: move this to compas_timber.fasteners
 from compas.data import Data
-from compas.geometry import Brep
-from compas.geometry import Cylinder
 from compas.geometry import Frame
 from compas.geometry import Line
 from compas.geometry import Transformation
@@ -69,19 +67,17 @@ class Fastener(TimberElement):
         # type: () -> int | None
         return self.graph_node
 
-    # TODO: implement Data instead of re-implementing
-    # def copy(self):
-    #     cls = type(self)
-    #     fast = cls(shape=self._shape, frame=self.frame)
-    #     fast.interfaces = [interface.copy() for interface in self.interfaces]
-    #     fast.debug_info = self.debug_info
-    #     return fast
+    @property
+    def __data__(self):
+        return {
+            "shape": self._shape,
+            "frame": self.frame,
+            "interfaces": self.interfaces,
+        }
 
-    # TODO: should implement compute_geometry instead
-    # @property
-    # def geometry(self):
-    #     """returns the geometry of the fastener in the model"""
-    #     return self.shape.transformed(Transformation.from_frame(self.frame))
+    def compute_geometry(self):
+        """returns the geometry of the fastener in the model"""
+        return self.shape.transformed(Transformation.from_frame(self.frame))
 
 
 class FastenerTimberInterface(Data):
@@ -95,8 +91,8 @@ class FastenerTimberInterface(Data):
 
     Parameters
     ----------
-    outline : :class:`~compas.geometry.Geometry`
-        The outline of the fastener geometry.
+    outline_points : List of :class:`~compas.geometry.Point`
+        The points of the polyline outline of the fastener geometry.
     thickness : float
         The thickness of the fastener plate.
     holes : list of dict, optional
@@ -116,8 +112,8 @@ class FastenerTimberInterface(Data):
 
     Attributes
     ----------
-    outline : :class:`~compas.geometry.Geometry`
-        The outline of the fastener geometry.
+    outline_points : List of :class:`~compas.geometry.Point`
+        The points of the polyline outline of the fastener geometry.
     thickness : float
         The thickness of the fastener plate.
     holes : list of dict, optional
@@ -136,18 +132,18 @@ class FastenerTimberInterface(Data):
 
     """
 
-    def __init__(self, outline, thickness, shapes=None, holes=None, frame=None, features=None):
+    def __init__(
+        self, outline_points=None, thickness=None, holes=None, shapes=None, frame=None, element=None, features=None
+    ):
         super(FastenerTimberInterface, self).__init__()
-        assert outline
-        assert thickness
-        self.outline = outline
+        self.outline_points = outline_points
         self.thickness = thickness
         self.holes = holes or []
         self.frame = frame or Frame.worldXY()
-        self.shapes = shapes
+        self.element = element
+        self.shapes = shapes or []
         self.features = features or []  # TODO: what are these? FeatureDefinitions?
         self._shape = None
-        self.test = []
 
     def __str__(self):
         return "FastenerTimberInterface at {}".format(self.frame)
@@ -155,62 +151,30 @@ class FastenerTimberInterface(Data):
     @property
     def __data__(self):
         return {
-            "outline": self.outline,
+            "outline_points": self.outline_points,
             "thickness": self.thickness,
             "holes": self.holes,
             "frame": self.frame,
+            "element": self.element,
             "shapes": self.shapes,
             "features": self.features,
         }
 
-    @property
-    def plate(self):
-        """Generate a plate from outline, thickness, and holes."""
-        if not self.outline:
-            return None
-        plate = Brep.from_extrusion(self.outline, Vector(0.0, 0.0, 1.0) * self.thickness)
-        for hole in self.holes:
-            frame = Frame(hole["point"], self.frame.xaxis, self.frame.yaxis)
-            hole = Brep.from_cylinder(Cylinder(hole["diameter"] / 2, self.thickness * 2, frame))
-            plate -= hole
-        return plate
-
-    @property
-    def shape(self):
-        """Return a Brep representation of the interface located at the WorldXY origin."""
-        if not self._shape:
-            geometries = []
-            if self.plate:
-                geometries.append(self.plate)
-            if self.shapes:
-                geometries.extend(self.shapes)
-
-            self._shape = geometries[0]
-            for geometry in geometries[1:]:
-                self._shape += geometry
-        return self._shape
-
-    @property
-    def geometry(self):
-        """returns the geometry of the interface in the model (oriented on the timber element)"""
-        return self.shape.transformed(Transformation.from_frame(self.frame))
-
-    def add_features(self, element):
+    def get_features(self, element):
         """Add a feature to the interface."""
         features = []
         for hole in self.holes:
             features.append(self._get_hole_feature(hole, element))
         # TODO: this uses the obsolete Feature classes, we should replace these with deffered BTLx
         for feature in self.features:
-            feature = feature.copy()
-            feature.transform(Transformation.from_frame(self.frame))
-            features.append(feature)
-        for feature in features:
-            element.add_feature(feature)
+            feat = feature.copy()
+            feat.transform(Transformation.from_frame(self.frame))
+            features.append(feat)
+        return features
 
     def _get_hole_feature(self, hole, element):
         """Get the line that goes through the timber element."""
-        vector = hole["vector"] or Vector(0.0, 0.0, 1.0)
+        vector = hole.get("vector", None) or Vector(0.0, 0.0, 1.0)
         length = vector.length
         point = hole["point"] - vector * length * 0.5
         drill_line = Line.from_point_direction_length(point, vector, length)